Create folder with ODMAP structure

This function creates an empty folder following the ODMAP structure (Overview, Data, Modeling, Assessment, and Prediction). It is intended to be used for habitat, movement, and connectivity modeling and suggested to be used by species (i.e. one folder per species), but this can be used in other ways depending of the user preferences.

Usage

create_odmap(path = ".", overwrite = FALSE)

Arguments

path

[character]
Path where the folder will be created, including intended folder name. If no argument is provided, the structure is created in the current working directory,

overwrite

[logical(1)=FALSE]
Should the folder be overwritten if' it already exists? Default is FALSE.

Examples

create_odmap(path = "./test")
#> ✔ Creating ./test/.
#> ✔ Creating ./test/01_overview/.
#> ! New project "02_data" is nested inside an existing project ./test/, which is
#>   rarely a good idea.
#> ℹ If this is unexpected, the here package has a function, `here::dr_here()`
#>   that reveals why ./test/ is regarded as a project.
#> Error in ui_yep(x = x, yes = yes, no = no, n_yes = n_yes, n_no = n_no,     shuffle = shuffle, .envir = .envir): ✖ User input required, but session is not interactive.
#> ℹ Query: "Do you want to create anyway?"