Get information from a raster

Get information from a raster

Usage

grass_raster_info(map)

Arguments

map

[character]
Name of the raster layer, possibly including the mapset in the format "map_name@mapset_name" if in a mapset different from the current mapset.

Value

A list with 9 elements, extracted from r.info: the four coorinates defining the bounding box of the raster (n, s, e, w); the raster resolution, taken as nsres; the number of rows, cols, and cells, and the type of data (e.g. CELL, FCELL).

Examples

# connect to GRASS
NinaR::grassConnect()
#> Error in loadNamespace(x): there is no package called ‘NinaR’
# get raster parameters
grass_raster_info(map = "master_grid_100m_norway@p_sam_tools")
#> Warning: error in running command
#> Error in parseGRASS(cmd, legacyExec = legacyExec): The command
#>    r.info --interface-description
#> could not be run (127), and produced the error message:
#>    sh: 1: r.info: not found